+++ src/cpu/memory/memory_core.rs
@@ -5,12 +5,12 @@ use std::fmt::Debug;
 use anyhow::Result;
 
 pub trait Memory: Debug {
-    fn read_mem_u8(&mut self, addr: u32) -> Result<u8>;
-    fn read_mem_u32(&mut self, addr: u32) -> Result<u32>;
-    fn read_mem_u16(&mut self, addr: u32) -> Result<u16>;
-    fn write_mem_u8(&mut self, addr: u32, value: u8) -> Result<()>;
-    fn write_mem_u16(&mut self, addr: u32, value: u16) -> Result<()>;
-    fn write_mem_u32(&mut self, addr: u32, value: u32) -> Result<()>;
-    fn read_buf(&mut self, addr: u32, buf: &mut [u8]) -> Result<()>;
-    fn write_buf(&mut self, addr: u32, buf: &[u8]) -> Result<()>;
+    fn read_mem_u8(&mut self, addr: u64) -> Result<u8>;
+    fn read_mem_u32(&mut self, addr: u64) -> Result<u32>;
+    fn read_mem_u16(&mut self, addr: u64) -> Result<u16>;
+    fn write_mem_u8(&mut self, addr: u64, value: u8) -> Result<()>;
+    fn write_mem_u16(&mut self, addr: u64, value: u16) -> Result<()>;
+    fn write_mem_u32(&mut self, addr: u64, value: u32) -> Result<()>;
+    fn read_buf(&mut self, addr: u64, buf: &mut [u8]) -> Result<()>;
+    fn write_buf(&mut self, addr: u64, buf: &[u8]) -> Result<()>;
 }
